1.) added cron and task management tests 2.) moved the core suite of tests into it's own file "originaltests".jl
