+++ src/utilities/fileutil.jl
@@ -1,5 +1,4 @@
 using HTTP
-#using Suppressor
 
 export readfile, mountedfolders, mountfolder
