+++ test/instancetests.jl
@@ -26,7 +26,6 @@ app2.get("/add/{a}/{b}") do req, a::Int, b::Int
     Dict("answer" => a + b) |> json
 end
 
- 
 # start both servers together
 app1.serve(port=8001, async=true, show_errors=false)
 app2.serve(port=8002, async=true, show_errors=false)
@@ -40,6 +39,31 @@ app2.serve(port=8002, async=true, show_errors=false)
     r = app2.internalrequest(HTTP.Request("GET", "/"))
     @test r.status == 200
     @test text(r) == "welcome to server #2"
+
+end
+
+
+@testset "testing add and subtract endpoints" begin
+
+    # Test subtract endpoint on server #1
+    r = app1.internalrequest(HTTP.Request("GET", "/subtract/10/5"))
+    @test r.status == 200
+    @test json(r)["answer"] == 5
+
+    # Test add endpoint on server #2
+    r = app2.internalrequest(HTTP.Request("GET", "/add/10/5"))
+    @test r.status == 200
+    @test json(r)["answer"] == 15
+
+    # Test subtract endpoint with negative result on server #1
+    r = app1.internalrequest(HTTP.Request("GET", "/subtract/5/10"))
+    @test r.status == 200
+    @test json(r)["answer"] == -5
+
+    # Test add endpoint with negative numbers on server #2
+    r = app2.internalrequest(HTTP.Request("GET", "/add/-10/-5"))
+    @test r.status == 200
+    @test json(r)["answer"] == -15
 
 end
